Conflicts:
	libgraphics/src/legacyblendmodes.cpp
	libgraphics/src/legacygradients.cpp

+++ toolchain/lc-compile/src/outputfile.h
@@ -23,7 +23,7 @@
 extern "C" {
 #endif
 
-typedef long PositionRef;
+typedef intptr_t PositionRef;
 typedef struct File *FileRef;
 
 enum DependencyModeType
